chore: media hooks 폴더 관리(#128)
